Refactor migration utils to be reused
