+++ pkg/apis/dashboard/migration/schemaversion/v40_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/grafana/grafana/pkg/apis/dashboard/migration/schemaversion"
-	"github.com/stretchr/testify/require"
 )
 
 func TestV40(t *testing.T) {
@@ -49,22 +48,4 @@ func TestV40(t *testing.T) {
 	}
 
 	runMigrationTests(t, tests, schemaversion.V40)
-}
-
-type migrationTestCase struct {
-	name     string
-	input    map[string]interface{}
-	expected map[string]interface{}
-}
-
-func runMigrationTests(t *testing.T, testCases []migrationTestCase, migrationFunc schemaversion.SchemaVersionMigrationFunc) {
-	t.Helper()
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			err := migrationFunc(tt.input)
-			require.NoError(t, err)
-			require.Equal(t, tt.expected, tt.input)
-		})
-	}
 }
